LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-transition-eval-expected.txt



FAIL Transition when the result of anchor() changes assert_equals: expected 110 but got 55
FAIL Transition when the result of anchor-size() changes assert_equals: expected 50 but got 0
